+++ app/seed-data/data/student.csv
@@ -1,7 +1,7 @@
-_id,fullname,email,gender,birthday
-students000000000000001,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2001
-students000000000000002,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2002
-students000000000000003,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2003
-students000000000000004,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2004
-students000000000000005,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2005
-students000000000000006,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2006
+_id,fullname,email,gender,birthday,code
+students000000000000001,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2001,1
+students000000000000002,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2002,2
+students000000000000003,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2003,3
+students000000000000004,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2004,4
+students000000000000005,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2005,5
+students000000000000006,Nguy?n V?n A,nguyenvana@gmail.com,0,01/01/2006,6
