+++ deploy/dev/.env.dev
@@ -4,11 +4,11 @@ SERVER_PORT=2999
 MONGODB_URL=mongo
 MONGODB_USER=root
 MONGODB_PASS=root
-MONGODB_DATABASE=contribution-management
+MONGODB_DATABASE=cm
 API_PREFIX=/api
 MONGODB_EXPRESS_PORT=4545
 MONGODB_PORT=22222
-CHECK_AUTH=true
+CHECK_AUTH=false
 CHECK_CHANGE_PASSWORD=false
 CHECK_REQUEST_SIGNATURE=false
 CHECK_BLOCKCHAIN=false
@@ -18,7 +18,7 @@ SEED_DATA=true
 URL_PRODUCTION=http://13.251.125.228:9001/
 PREFIX_BLOCKCHAIN_ID=1.0
 VERSION=1.0
-MEMORY_LIMIT=3072M
+MEMORY_LIMIT=1000M
 MEMORY_RESERVATIONS=128M
 TOTAL_QR=71
 NUMBER_QR_PER_PAGE=35
